add more reducers
