+++ src/reducers/index.js
@@ -1,14 +1,10 @@
-import charactersJson from '../data/characters.json';
-import { ADD_CHARACTER } from '../actions/index'
+import { combineReducers } from 'redux';
+import character from './character';
+import hero from './hero';
 
-const characters = (state = charactersJson, action) => {
-  switch  (action.type) {
-    case ADD_CHARACTER:
-      let character = state.filter(item => item.id !== action.id);
-      return character;
-    default:
-      return state;
-  }
-}
+const rootReducer = combineReducers({
+  character,
+  hero
+})
 
-export default characters;
+export default rootReducer;
